refactor(settings): split non-client types from client hook; fix UI copy
- Move DEADLINE_OPTIONS, DeadlineDays, UserSettings, DEFAULT_SETTINGS,
  and isDeadlineDays into settingsTypes.ts (no 'use client') so they are
  safe to import from server components and shared code. useSettings.ts
  now imports from there and stays the only client module.
- Reword deadline dropdown description to "you'd like to be alerted".
