+++ frontend/src/lib/useSettings.ts
@@ -1,28 +1,10 @@
 'use client';
 
 import { useState } from 'react';
-
-export const DEADLINE_OPTIONS = [1, 2, 3, 5, 7, 14] as const;
-export type DeadlineDays = (typeof DEADLINE_OPTIONS)[number];
-
-export type UserSettings = {
-  notificationsEnabled: boolean;
-  daysBeforeDeadline: DeadlineDays;
-  googleConnected: boolean;
-};
-
-const DEFAULTS: UserSettings = {
-  notificationsEnabled: true,
-  daysBeforeDeadline: 3,
-  googleConnected: false,
-};
-
-export function isDeadlineDays(value: number): value is DeadlineDays {
-  return (DEADLINE_OPTIONS as readonly number[]).includes(value);
-}
+import { DEFAULT_SETTINGS, DeadlineDays, UserSettings } from './settingsTypes';
 
 export function useSettings() {
-  const [settings, setSettings] = useState<UserSettings>(DEFAULTS);
+  const [settings, setSettings] = useState<UserSettings>(DEFAULT_SETTINGS);
 
   return {
     settings,
